requirements.txt
alembic==1.6.5
asyncpg==0.25.0
fastapi==0.75.0
gunicorn==20.1.0
httpx==0.18.1
Jinja2==3.0.3
ormar==0.10.25
psycopg2-binary==2.9.3
pydantic==1.9.0
python-multipart==0.0.5
python-dotenv==0.19.2
SQLAlchemy==1.4.31
tenacity==8.0.1
uvicorn==0.17.6
uvloop==0.16.0
